# backend/postman/specs/openapi.yaml
openapi: 3.0.3
info:
  title: MSM — Sistema MIP de Productividad y Reconocimiento Comercial
  description: |
    API del sistema MSM (Sistema MIP) para medir, gestionar y reconocer el
    desempeño de la fuerza de ventas farmacéutica.

    **Autenticación**: Bearer JWT. Usar `POST /auth/login` con form-data
    (`username` + `password`) para obtener el `access_token`.

    **Roles**: ADMIN · PRESIDENCIA · DIR_COMERCIAL · GERENTE_PRODUCTIVIDAD ·
    GERENTE_DISTRITO · GERENTE_MARCA · REPRESENTANTE_MEDICO · CONSULTA
  version: 1.0.0
  contact:
    email: segura.moises@gmail.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RankingItem:
      type: object
      properties:
        posicion:        { type: integer }
        posicion_linea:  { type: integer, nullable: true }
        variacion:       { type: number, nullable: true }
        rm_id:           { type: integer }
        rm_codigo:       { type: string }
        rm_nombre:       { type: string }
        linea_nombre:    { type: string }
        gerente_nombre:  { type: string }
        score_total:     { type: number }
        score_ciclo_actual: { type: number }
        num_ciclos:      { type: integer }
        elegible:        { type: boolean }
        tipo_ranking:    { type: string }
        ciclo_efectivo:  { type: integer, nullable: true }
...
security:
  - BearerAuth: []
paths:
...
  /ranking:
    get:
      tags: [Ranking]
      summary: Ranking acumulado paginado
      description: |
        Calcula el score promedio acumulado de los ciclos 1..N del año seleccionado.
        Los RMs con rol REPRESENTANTE_MEDICO solo ven su propio registro.
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
        - name: tipo
          in: query
          schema: { type: string, default: MENSUAL, enum: [MENSUAL, TRIMESTRAL, ANUAL, REGIONAL] }
        - name: top
          in: query
          schema: { type: integer, description: Limitar a top N resultados }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Ranking paginado
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/RankingItem' }
                  total:           { type: integer }
                  page:            { type: integer }
                  size:            { type: integer }
                  pages:           { type: integer }
                  ciclo_efectivo:  { type: integer, nullable: true }
                  ciclos_numeros:  { type: array, items: { type: integer } }
  /ranking/regional:
    get:
      tags: [Ranking]
      summary: Ranking regional multipaís
      description: Solo ADMIN, PRESIDENCIA, DIR_COMERCIAL, GERENTE_PRODUCTIVIDAD
      parameters:
        - name: ciclo_id
          in: query
          schema: { type: integer }
        - name: top
          in: query
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Top RMs multipaís
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
        '403': { description: Sin permiso }
  /ranking/anual:
    get:
      tags: [Ranking]
      summary: Ranking anual histórico
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Ranking anual
          content:
            application/json:
              schema:
                type: array
                items: { type: object }
  /ranking/generar:
    post:
      tags: [Ranking]
      summary: Disparar generación de ranking en background
      description: Solo ADMIN, GERENTE_PRODUCTIVIDAD
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pais_id, ciclo_id, tipo_ranking]
              properties:
                pais_id:      { type: integer }
                ciclo_id:     { type: integer }
                tipo_ranking: { type: string, enum: [MENSUAL, TRIMESTRAL, ANUAL, REGIONAL] }
      responses:
        '200':
          description: Cálculo iniciado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:  { type: string }
                  pais_id:  { type: integer }
                  ciclo_id: { type: integer }

  # ─── DASHBOARD ────────────────────────────────────────────────────────────
...
  /dashboard/productividad:
    get:
      tags: [Dashboard]
      summary: Dashboard de Productividad (cobertura, promedio diario, farmacias)
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
      responses:
        '200': { description: KPIs de productividad }
...
  /productividad:
    get:
      tags: [Productividad]
      summary: KPIs de productividad paginados
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
        - name: linea_id
          in: query
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200': { description: KPIs paginados }
  /productividad/rm/{rm_id}:
    get:
      tags: [Productividad]
      summary: KPIs de un RM específico
      parameters:
        - name: rm_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: KPIs del RM }
        '404': { description: RM no encontrado }
  /productividad/pais/{id}:
    get:
      tags: [Productividad]
      summary: KPIs agregados por país
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: KPIs del país }
  /productividad/resumen:
    get:
      tags: [Productividad]
      summary: Resumen consolidado de productividad
      parameters:
        - name: pais_id
          in: query
          schema: { type: integer }
        - name: ciclo_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Consolidado con total_rms, cumplimiento_promedio y estadísticas de puntaje

  # ─── COMERCIAL ────────────────────────────────────────────────────────────
...